defined strategies for DeleteTableStorage and EqualityDelete
