+++ deltacat/tests/compute/test_compact_partition_rebase_then_incremental.py
@@ -7,13 +7,6 @@ from boto3.resources.base import ServiceResource
 import pyarrow as pa
 from deltacat.io.ray_plasma_object_store import RayPlasmaObjectStore
 from pytest_benchmark.fixture import BenchmarkFixture
-from deltacat.compute.compactor_v2.deletes.model import DeleteStrategy
-from deltacat.compute.compactor_v2.deletes.noop_delete_strategy import (
-    NOOPDeleteStrategy,
-)
-from deltacat.compute.compactor_v2.deletes.default_equality_delete_strategy import (
-    DefaultEqualityDeleteStrategy,
-)
 
 from deltacat.tests.compute.test_util_constant import (
     BASE_TEST_SOURCE_NAMESPACE,
